Divisione tra interi da interi se non viene fatto il casting,eseguito casting

+++ gdb/gdb1.c
@@ -23,7 +23,7 @@ int main()
 
 	double media;
 	if (numeri > 0) {
-		media = somma/numeri;
+		media = (float)somma/numeri;
 		printf("La media è %f\n", media);
 	}
